refactor(#223): Home에서 사용하는 컴포넌트를 /home/components/로 이동
- Home에서 사용하는 컴포넌트들을 /home/components/로 이동
- SideBar에서만 사용하는 컴포넌트들은 /home/components/sidebar/로 이동
- 사용하지 않는 CarDetailData.json 더미데이터 삭제
