--- client/src/pages/cars/CarDetailData.json
@@ -1,40 +1,0 @@
-{
-  "success": true,
-  "code": 200,
-  "message": "차량 상세 정보 조회 성공",
-  "data": {
-    "car": {
-      "carName": "제네시스 G80",
-      "carNumber": "00가0000",
-      "imageUrls": [
-        "/car3.jpg",
-        "/car1.jpg",
-        "/car-inside.jpg",
-        "/car-detail-front.jpg",
-        "/car-side.jpg"
-      ],
-      "mileage": 10,
-      "fuel": "전기",
-      "type": "승용차",
-      "capacity": 4,
-      "year": 2010,
-      "feePerHour": 10000,
-      "carAddress": "경기도 오산시 밀머리로",
-      "description": "안녕하세요 저는 호스트입니다. 빌리면 소중히 사용하세요.\n저렴한 가격에 빌려가세요. 난폭운전 금지 부탁드립니다.\n위치는 예약 확정하면 알려드림\n안녕하세요 저는 호스트입니다. 빌리면 소중히 사용하세요.\n저렴한 가격에 빌려가세요. 난폭운전 금지 부탁드립니다.\n위치는 예약 확정하면 알려드림",
-      "dates": [
-        [
-          "2024-02-06T00:00:00",
-          "2024-02-13T23:59:59"
-        ],
-        [
-          "2024-02-20T00:00:00",
-          "2024-02-29T23:59:59"
-        ]
-      ],
-      "host": {
-        "name": "오수현",
-        "profileImg": "/logo.png"
-      }
-    }
-  }
-}
