+++ client/src/pages/home/components/CarCard.tsx
